be/requirements.txt
asgiref==3.10.0
Django==5.2.7
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
dotenv==0.9.9
gunicorn==23.0.0
psycopg2==2.9.10